pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macd-regime-report"
version = "0.1.0"
description = "Rule-driven MACD regime report engine with state machine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "PyYAML>=6.0",
  "yfinance>=0.2.54",
  "pandas-datareader>=0.10.0",
  "tabulate>=0.9.0",
]

[project.scripts]
macd-regime = "macd_regime.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
